Check for valid game storage path
Added a while loop that checks if the game storage path is valid. If no exceptions are raised, then it will break out of the while loop.

It also checks for the settings.txt file if the path is valid.

If an error is raised, it will specify what it can't find and ask the user to try again.